requirements.txt
O365>=2.0.0
openai-whisper>=20231117
python-dotenv>=1.0.0
APScheduler>=3.10.0
